src/eda/pyproject.toml
[project]
name = "eda"
version = "0.1.0"
description = "FabCon 2026 session analysis — preprocess, classify, cluster, reduce, model-select"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # Core CLI
  "click>=8.1.0",
  "rich>=13.7.0",
  # Data
  "pandas>=2.2.0",
  "numpy>=2.0.0",
  "python-frontmatter>=1.1.0",
  # ML
  "scikit-learn>=1.6.0",
  "joblib>=1.4.0",
  # Scientific
  "scipy>=1.14.1",
  # Visualization
  "matplotlib>=3.9.3",
  # I/O (pandas optional deps)
  "pyarrow>=13.0.0",        # parquet/feather
  "openpyxl>=3.1.5",        # Excel write
  "tabulate>=0.9.0",        # table formatting
  # Performance
  "bottleneck>=1.4.2",      # fast NaN aggregations
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5.6"]
hdf5 = ["tables>=3.10.1"]
sql = ["SQLAlchemy>=2.0.36"]
excel-full = ["xlrd>=2.0.1", "xlsxwriter>=3.2.0"]
parquet-alt = ["fastparquet>=2024.11.0"]
aws = ["s3fs>=2024.10.0"]
gcp = ["gcsfs>=2024.10.0"]
xml = ["lxml>=5.3.0"]
html = ["beautifulsoup4>=4.12.3", "lxml>=5.3.0"]
performance = ["numba>=0.60.0", "numexpr>=2.10.2"]
all = [
  "umap-learn>=0.5.6",
  "tables>=3.10.1",
  "SQLAlchemy>=2.0.36",
  "xlrd>=2.0.1",
  "xlsxwriter>=3.2.0",
  "fastparquet>=2024.11.0",
  "s3fs>=2024.10.0",
  "lxml>=5.3.0",
  "beautifulsoup4>=4.12.3",
  "numba>=0.60.0",
  "numexpr>=2.10.2",
]

[project.scripts]
eda = "eda.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eda"]
